% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Age_Analysis.R
\name{Age_Analysis}
\alias{Age_Analysis}
\title{processes age data in R and creates output. Results in Age distribution curve,
Categorized the age, and bar graph. Returns Age_cat as a value}
\usage{
Age_Analysis(data, Age)
}
\arguments{
\item{data}{is data.}

\item{Age}{is age.}
}
\description{
processes age data in R and creates output. Results in Age distribution curve,
Categorized the age, and bar graph. Returns Age_cat as a value
}
\examples{
Age_Analysis(data, Age)
}
